CGMES metadata model interface and implementation
