+++ cgmes/cgmes-extensions/src/main/java/com/powsybl/cgmes/extensions/CgmesMetadataModelsAdderImpl.java
@@ -8,6 +8,7 @@
 package com.powsybl.cgmes.extensions;
 
 import com.powsybl.cgmes.model.CgmesMetadataModel;
+import com.powsybl.cgmes.model.CgmesMetadataModelImpl;
 import com.powsybl.cgmes.model.CgmesSubset;
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.commons.extensions.AbstractExtensionAdder;
@@ -93,7 +94,7 @@ class CgmesMetadataModelsAdderImpl extends AbstractExtensionAdder<Network, Cgmes
             if (profiles.isEmpty()) {
                 throw new PowsyblException("Model must contain at least one profile");
             }
-            models.add(new CgmesMetadataModel(subset, modelingAuthoritySet)
+            models.add(new CgmesMetadataModelImpl(subset, modelingAuthoritySet)
                     .setId(id)
                     .setDescription(description)
                     .setVersion(version)
